backend/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
sqlalchemy==2.0.21
psycopg2-binary==2.9.7
pydantic==2.4.2
python-dotenv==1.0.0
boto3==1.28.57